requirements.txt
Flask>=2.0
python-dotenv>=0.19
google-generativeai>=0.5.0
moviepy==1.0.3
requests>=2.25.0
Pillow>=9.0.0
scipy>=1.0.0